feat: carry on (#16)
Waow!

See
https://github.com/terrarium-modpack/ShutUpAnnoyingMod/commit/2f0bddb9bd6425a42b582292fb0fea24dca137f1
for more details on the first 2 settings.

Basically, by default, it actually serializes the NBT of the block
you're holding to a string every tick.
The scripting feature is kind of broken - in theory, someone with
creative could abuse it (even if it's off in the config) to execute
commands as console. Not good.